Refactor EigenSparseMarker and RaggedSparseMarker
They were previously in sparsemarker.h and sparsemarker.cpp
